+++ lib/app.ts
@@ -1,5 +1,7 @@
 import { Channel } from './channel'
 import { AppConfigAPI, AppState } from './types'
+import { OnConfigureHandler } from './types/app.types'
+import { KeyValueMap } from './types/entities'
 
 const HOOK_STAGE_PRE_INSTALL = 'preInstall'
 const HOOK_STAGE_POST_INSTALL = 'postInstall'
@@ -97,16 +99,16 @@ export default function createApp(channel: Channel): AppConfigAPI {
     isInstalled() {
       return channel.call('callAppMethod', 'isInstalled') as Promise<boolean>
     },
-    getParameters() {
-      return channel.call('callAppMethod', 'getParameters') as Promise<any>
+    getParameters<T extends KeyValueMap = KeyValueMap>() {
+      return channel.call('callAppMethod', 'getParameters') as Promise<T | null>
     },
     getCurrentState() {
       return channel.call('callAppMethod', 'getCurrentState') as Promise<AppState | null>
     },
-    onConfigure(handler: Function) {
+    onConfigure(handler: OnConfigureHandler) {
       setHandler(HOOK_STAGE_PRE_INSTALL, handler)
     },
-    onConfigurationCompleted(handler: Function) {
+    onConfigurationCompleted(handler: (err: null | { message: string }) => void) {
       setHandler(HOOK_STAGE_POST_INSTALL, handler)
     },
   }