+++ CHANGELOG.md
@@ -1,3 +1,117 @@
+# [4.0.0-alpha.18](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.17...v4.0.0-alpha.18) (2021-06-23)
+
+
+### Features
+
+* add type parameter to create task input data ([#807](https://github.com/contentful/ui-extensions-sdk/issues/807)) ([493bb0e](https://github.com/contentful/ui-extensions-sdk/commit/493bb0ec92969263ea549e6f6cc1c932c33a60c6))
+
+# [4.0.0-alpha.17](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.16...v4.0.0-alpha.17) (2021-06-21)
+
+
+### Features
+
+* add interface for space.getTeams to v4 [HOMER-84] ([#797](https://github.com/contentful/ui-extensions-sdk/issues/797)) ([0cfe659](https://github.com/contentful/ui-extensions-sdk/commit/0cfe659970233fdc90b62d799cd608d949f7e8f7))
+
+# [4.0.0-alpha.16](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.15...v4.0.0-alpha.16) (2021-06-15)
+
+
+### Bug Fixes
+
+* set ContentTypeAPI in SharedEditorSDK ([#794](https://github.com/contentful/ui-extensions-sdk/issues/794)) ([b9fa7fc](https://github.com/contentful/ui-extensions-sdk/commit/b9fa7fc59fb23aa691f00401a9b0f8bbd57bcba2))
+
+# [4.0.0-alpha.15](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.14...v4.0.0-alpha.15) (2021-06-15)
+
+
+### Bug Fixes
+
+* export ContentTypeAPI ([#793](https://github.com/contentful/ui-extensions-sdk/issues/793)) ([2f0f74e](https://github.com/contentful/ui-extensions-sdk/commit/2f0f74e7674604de88f3510343949885b333bdcd))
+
+# [4.0.0-alpha.14](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.13...v4.0.0-alpha.14) (2021-06-15)
+
+
+### Bug Fixes
+
+* use correct typings in waitUntilAssetProcessed ([#792](https://github.com/contentful/ui-extensions-sdk/issues/792)) ([74a030a](https://github.com/contentful/ui-extensions-sdk/commit/74a030ab4351fccc5c3ff28f7a6d4d51e368521a))
+
+# [4.0.0-alpha.13](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.12...v4.0.0-alpha.13) (2021-06-14)
+
+
+### Features
+
+* add due date as a task field type ([#783](https://github.com/contentful/ui-extensions-sdk/issues/783)) ([f2b216a](https://github.com/contentful/ui-extensions-sdk/commit/f2b216a6b5906e8b2a1a0c338bb4bf9504612e9b))
+
+# [4.0.0-alpha.12](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.11...v4.0.0-alpha.12) (2021-06-14)
+
+
+### Bug Fixes
+
+* change ContentType API to public type ([#788](https://github.com/contentful/ui-extensions-sdk/issues/788)) ([88742a1](https://github.com/contentful/ui-extensions-sdk/commit/88742a18b363840465a3f229f5d859af0682c0c0))
+
+# [4.0.0-alpha.11](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.10...v4.0.0-alpha.11) (2021-06-04)
+
+
+### Bug Fixes
+
+* update onConfigure type to match reference ([#765](https://github.com/contentful/ui-extensions-sdk/issues/765)) ([2819f13](https://github.com/contentful/ui-extensions-sdk/commit/2819f1325c24c5cb210d5dd071dfbef1ae875ac0))
+
+# [4.0.0-alpha.10](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.9...v4.0.0-alpha.10) (2021-05-31)
+
+
+### Features
+
+* update AppConfigAPI types ([#753](https://github.com/contentful/ui-extensions-sdk/issues/753)) ([6876d0a](https://github.com/contentful/ui-extensions-sdk/commit/6876d0ad0f8214acdfb97783caa54740410b037d))
+
+# [4.0.0-alpha.9](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.8...v4.0.0-alpha.9) (2021-05-28)
+
+
+### Features
+
+* remove files from dist that were kept for backwards compatibility ([#750](https://github.com/contentful/ui-extensions-sdk/issues/750)) ([dbed8c2](https://github.com/contentful/ui-extensions-sdk/commit/dbed8c29eb9bbfc45bc5c50944bed75e8b3616e1))
+
+# [4.0.0-alpha.8](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.7...v4.0.0-alpha.8) (2021-05-27)
+
+
+### Features
+
+* improve types for Field & Entry API ([#745](https://github.com/contentful/ui-extensions-sdk/issues/745)) ([4774448](https://github.com/contentful/ui-extensions-sdk/commit/4774448bd766b83fadfc7e989bee570ab068434c))
+
+# [4.0.0-alpha.7](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.6...v4.0.0-alpha.7) (2021-05-26)
+
+
+### Features
+
+* better typings for user API ([#744](https://github.com/contentful/ui-extensions-sdk/issues/744)) ([29a6359](https://github.com/contentful/ui-extensions-sdk/commit/29a63594933f1063ac8734a9877fec6fa25928b9))
+
+# [4.0.0-alpha.6](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.5...v4.0.0-alpha.6) (2021-05-20)
+
+
+### Bug Fixes
+
+* navigator api types ([#725](https://github.com/contentful/ui-extensions-sdk/issues/725)) ([7cac2eb](https://github.com/contentful/ui-extensions-sdk/commit/7cac2eb08200bba51d95d5461f00011a3add771b)), closes [#718](https://github.com/contentful/ui-extensions-sdk/issues/718) [#719](https://github.com/contentful/ui-extensions-sdk/issues/719) [#715](https://github.com/contentful/ui-extensions-sdk/issues/715) [#720](https://github.com/contentful/ui-extensions-sdk/issues/720) [#723](https://github.com/contentful/ui-extensions-sdk/issues/723) [#721](https://github.com/contentful/ui-extensions-sdk/issues/721)
+
+# [4.0.0-alpha.5](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.4...v4.0.0-alpha.5) (2021-05-18)
+
+
+### Features
+
+* expose WithId ([#722](https://github.com/contentful/ui-extensions-sdk/issues/722)) ([80189db](https://github.com/contentful/ui-extensions-sdk/commit/80189dbdf5b7592a73e1ed4ec4fef40b0c85a1c8))
+
+# [4.0.0-alpha.4](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.3...v4.0.0-alpha.4) (2021-05-18)
+
+
+### Bug Fixes
+
+* more explicit space api types ([#711](https://github.com/contentful/ui-extensions-sdk/issues/711)) ([42d63de](https://github.com/contentful/ui-extensions-sdk/commit/42d63dea5fd9a313918a9591311d716c3374ac60))
+
+
+# [3.38.0](https://github.com/contentful/ui-extensions-sdk/compare/v3.37.0...v3.38.0) (2021-06-30)
+
+
+### Features
+
+* add resolved info to tasks sys ([#822](https://github.com/contentful/ui-extensions-sdk/issues/822)) ([e0be824](https://github.com/contentful/ui-extensions-sdk/commit/e0be824956082aaa8fe00bd073f7354fb7dea51b))
+
+>>>>>>> canary
 # [3.37.0](https://github.com/contentful/ui-extensions-sdk/compare/v3.36.0...v3.37.0) (2021-06-23)
 
 
@@ -20,53 +134,77 @@
 * add due date as a task field type ([#776](https://github.com/contentful/ui-extensions-sdk/issues/776)) ([ed0bf60](https://github.com/contentful/ui-extensions-sdk/commit/ed0bf60ce877665a3558555346c6bf0c9274bb49))
 
 ## [3.34.3](https://github.com/contentful/ui-extensions-sdk/compare/v3.34.2...v3.34.3) (2021-05-18)
+>>>>>>> master
 
 
 ### Reverts
 
 * "chore: test that SDK respects alias ([#710](https://github.com/contentful/ui-extensions-sdk/issues/710))" ([#717](https://github.com/contentful/ui-extensions-sdk/issues/717)) ([4ab6e48](https://github.com/contentful/ui-extensions-sdk/commit/4ab6e48237ef6d4cdd75c7721cd994539ced0ce2))
 
-## [3.34.2](https://github.com/contentful/ui-extensions-sdk/compare/v3.34.1...v3.34.2) (2021-05-11)
+# [4.0.0-alpha.3](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.2...v4.0.0-alpha.3) (2021-05-11)
 
+### Bug Fixes
+
+- add missing types also in files ([#707](https://github.com/contentful/ui-extensions-sdk/issues/707)) ([957c858](https://github.com/contentful/ui-extensions-sdk/commit/957c85805a683832ca7f856fe525d17195a910d2))
+
+## [3.34.3](https://github.com/contentful/ui-extensions-sdk/compare/v3.34.2...v3.34.3) (2021-05-18)
 
 ### Reverts
 
-* bump cypress from 7.2.0 to 7.3.0 ([#706](https://github.com/contentful/ui-extensions-sdk/issues/706)) ([ebcdf01](https://github.com/contentful/ui-extensions-sdk/commit/ebcdf01c26c59707e5733f34d8897a9de9bfa462))
+- bump cypress from 7.2.0 to 7.3.0 ([#706](https://github.com/contentful/ui-extensions-sdk/issues/706)) ([ebcdf01](https://github.com/contentful/ui-extensions-sdk/commit/ebcdf01c26c59707e5733f34d8897a9de9bfa462))
+
+# [4.0.0-alpha.2](https://github.com/contentful/ui-extensions-sdk/compare/v4.0.0-alpha.1...v4.0.0-alpha.2) (2021-05-11)
+
+### Bug Fixes
+
+- include typings in the release ([#699](https://github.com/contentful/ui-extensions-sdk/issues/699)) ([16b15e2](https://github.com/contentful/ui-extensions-sdk/commit/16b15e25ddcfd8a34a6f4c9732fc5e67929ac62b))
+
+# [4.0.0-alpha.1](https://github.com/contentful/ui-extensions-sdk/compare/v3.34.1...v4.0.0-alpha.1) (2021-05-10)
+
+### Features
+
+- improve space API typings ([#684](https://github.com/contentful/ui-extensions-sdk/issues/684)) ([a57354c](https://github.com/contentful/ui-extensions-sdk/commit/a57354c32b78204d285aba897d213f201fec208c))
+
+### BREAKING CHANGES
+
+- old typings were too loose.
+- "chore: test that SDK respects alias ([#710](https://github.com/contentful/ui-extensions-sdk/issues/710))" ([#717](https://github.com/contentful/ui-extensions-sdk/issues/717)) ([4ab6e48](https://github.com/contentful/ui-extensions-sdk/commit/4ab6e48237ef6d4cdd75c7721cd994539ced0ce2))
+
+## [3.34.2](https://github.com/contentful/ui-extensions-sdk/compare/v3.34.1...v3.34.2) (2021-05-11)
+
+### Reverts
+
+- bump cypress from 7.2.0 to 7.3.0 ([#706](https://github.com/contentful/ui-extensions-sdk/issues/706)) ([ebcdf01](https://github.com/contentful/ui-extensions-sdk/commit/ebcdf01c26c59707e5733f34d8897a9de9bfa462))
 
 ## [3.34.1](https://github.com/contentful/ui-extensions-sdk/compare/v3.34.0...v3.34.1) (2021-05-06)
 
-
 ### Bug Fixes
 
-* make options of dialogs.openExtension optional ([#681](https://github.com/contentful/ui-extensions-sdk/issues/681)) ([ca6d15b](https://github.com/contentful/ui-extensions-sdk/commit/ca6d15bf6c9979410296dfaa3f2ad3bd71ba33c4))
+- make options of dialogs.openExtension optional ([#681](https://github.com/contentful/ui-extensions-sdk/issues/681)) ([ca6d15b](https://github.com/contentful/ui-extensions-sdk/commit/ca6d15bf6c9979410296dfaa3f2ad3bd71ba33c4))
 
 # [3.34.0](https://github.com/contentful/ui-extensions-sdk/compare/v3.33.0...v3.34.0) (2021-04-30)
 
-
 ### Features
 
-* add navigate to lists methods ([ef5ee41](https://github.com/contentful/ui-extensions-sdk/commit/ef5ee4148911f51b333dbd0967d5fba2748de317))
+- add navigate to lists methods ([ef5ee41](https://github.com/contentful/ui-extensions-sdk/commit/ef5ee4148911f51b333dbd0967d5fba2748de317))
 
 # [3.33.0](https://github.com/contentful/ui-extensions-sdk/compare/v3.32.2...v3.33.0) (2021-04-06)
 
-
 ### Features
 
-* expose tags api ([3ae6b4b](https://github.com/contentful/ui-extensions-sdk/commit/3ae6b4b45a16f04cb931218b6df654e9bf97cc1b))
+- expose tags api ([3ae6b4b](https://github.com/contentful/ui-extensions-sdk/commit/3ae6b4b45a16f04cb931218b6df654e9bf97cc1b))
 
 ## [3.32.2](https://github.com/contentful/ui-extensions-sdk/compare/v3.32.1...v3.32.2) (2021-03-23)
 
-
 ### Bug Fixes
 
-* [EXT-2508] use husky v4 ([#611](https://github.com/contentful/ui-extensions-sdk/issues/611)) ([85e1d4f](https://github.com/contentful/ui-extensions-sdk/commit/85e1d4f43e710b747f0ecba9d35396d8dfaacc10))
+- [EXT-2508] use husky v4 ([#611](https://github.com/contentful/ui-extensions-sdk/issues/611)) ([85e1d4f](https://github.com/contentful/ui-extensions-sdk/commit/85e1d4f43e710b747f0ecba9d35396d8dfaacc10))
 
 ## [3.32.1](https://github.com/contentful/ui-extensions-sdk/compare/v3.32.0...v3.32.1) (2021-02-16)
 
-
 ### Bug Fixes
 
-* link to App SDK reference ([#550](https://github.com/contentful/ui-extensions-sdk/issues/550)) ([be47635](https://github.com/contentful/ui-extensions-sdk/commit/be47635c3c5b621e501df3643292c6c2bd2ef417))
+- link to App SDK reference ([#550](https://github.com/contentful/ui-extensions-sdk/issues/550)) ([be47635](https://github.com/contentful/ui-extensions-sdk/commit/be47635c3c5b621e501df3643292c6c2bd2ef417))
 
 # [3.32.0](https://github.com/contentful/ui-extensions-sdk/compare/v3.31.1...v3.32.0) (2021-01-29)